requirements.txt
Django==1.8
django-bootstrap-form==3.2.1
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.8
psycopg2==2.6.2
requests==2.13.0
